Sample data from bowhead whales now in data frame format, and test functions updated to work directly with that format

+++ pkg/R/genepop.to.genind.R
@@ -3,20 +3,24 @@
 #' Imports a \code{.txt} file in \code{GenePop} format into an object of type \code{genind}
 
 #' 
-#' The main work is done by the function \code{adegenet::import2genind}. However, that function requires text files with an extension of \code{.gen}, whereas such files usually have extension \code{.txt}. The sole purpose of this function is to work around the \dQuote{.gen} requirement.
+#' The main work is done by the function \code{adegenet::read.genepop}. However, that function requires text files with an extension of \code{.gen}, whereas such files usually have extension \code{.txt}. The sole purpose of this function is to work around the \dQuote{.gen} requirement.
 #' 
 #' @param name the name of a file in \code{GenePop} format
 #' @param quiet whether a conversion message should be printed
+#' @param ncode Set to the number of characters per allele name
 #' 
 #' @return an object of class \code{genind}
 #' 
 #' @export
-genepop.to.genind <- function(name, quiet=TRUE){
-	require(adegenet)
-	tempfile <- file(name)
-	tmp <- readLines(tempfile)
-	writeLines(tmp, "tempgenepop.gen")
-	ind <- adegenet::import2genind("tempgenepop.gen", quiet=quiet)
-	unlink("tempgenepop.gen")
-	ind
+genepop.to.genind <- function(name, quiet = TRUE, ncode = 3) {
+	if (requireNamespace("adegenet")) {
+		tempfile <- file(name)
+		tmp <- readLines(tempfile)
+		writeLines(tmp, "tempgenepop.gen")
+		ind <- adegenet::read.genepop("tempgenepop.gen", quiet = quiet, ncode = ncode)
+		unlink("tempgenepop.gen")
+		return(ind)
+	} else {
+		return(0)
+	}
 }
